@@ -1,6 +1,6 @@
 module github.com/FerretDB/FerretDB/tools
 
-go 1.19
+go 1.20
 
 require (
 	github.com/BurntSushi/go-sumtype v0.0.0-20221020234012-480526a59796
