@@ -14,8 +14,8 @@
 
 // Keep both old and new styles of build tags.
 
-//go:build go1.19
-// +build go1.19
+//go:build go1.20
+// +build go1.20
 
 package tools
 
@@ -35,7 +35,7 @@ import (
 )
 
 // Check that `go` in $PATH have the right version.
-// Catches problems like `/some/path/go generate` invocations where `/some/path/go` is 1.19+
+// Catches problems like `/some/path/go generate` invocations where `/some/path/go` is 1.20+
 // (that's checked by the build tags above), but just `go` in $PATH (typically something like `/usr/bin/go`)
 // is an earlier version.
 
